Address Remaining Code Review Feedback
https://github.com/colszowka/simplecov/pull/694#pullrequestreview-256178924
ref #781
