+++ spec/combine/results_combiner_spec.rb
@@ -2,7 +2,7 @@
 
 require "helper"
 
-describe SimpleCov::RunResultsCombiner do
+describe SimpleCov::Combine::ResultsCombiner do
   describe "with two faked coverage resultsets" do
     before do
       @resultset1 = {
@@ -39,7 +39,7 @@ describe SimpleCov::RunResultsCombiner do
 
     context "a merge" do
       subject do
-        SimpleCov::RunResultsCombiner.combine(@resultset1, @resultset2, @resultset3)
+        SimpleCov::Combine::ResultsCombiner.combine(@resultset1, @resultset2, @resultset3)
       end
 
       it "has proper results for sample.rb" do
@@ -93,7 +93,7 @@ describe SimpleCov::RunResultsCombiner do
       source_fixture("sample.rb").freeze => {:lines => [1, nil, 1, 1, nil, nil, 1, 1, nil, nil]}
     }
 
-    merged_result = SimpleCov::RunResultsCombiner.combine(resultset1, resultset2)
+    merged_result = SimpleCov::Combine::ResultsCombiner.combine(resultset1, resultset2)
     expect(merged_result.keys).to eq(resultset1.keys)
     expect(merged_result.values.map(&:frozen?)).to eq([false, false])
